requirements.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1

# Authentication and security
bcrypt==4.1.2
authlib==1.3.0

# AI/ML dependencies
groq>=1.1.1
sentence-transformers>=2.7.0

# Web scraping and HTTP
requests==2.31.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2

# Database & Vector Support
psycopg2-binary>=2.9.9
pgvector>=0.2.5

# Additional ML/Data Science
pandas>=1.1
numpy>=1.19
scikit-learn>=1.0
scipy>=1.7

# Production server
gunicorn>=20.1.0
